chore/Rename azure folder to azure_services
