+++ django/azure_services/adapters.py
