+++ django/azure_services/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class AzureConfig(AppConfig):
+    name = 'azure_services'
