+++ django/azure_services/provider.py
