+++ django/azure_services/tasks.py
