+++ django/azure_services/urls.py
