+++ django/azure_services/views.py
