+++ django/tiip/settings.py
@@ -5,6 +5,8 @@ import sys
 from environs import Env
 from django.utils.translation import ugettext_lazy as _
 
+from azure.monitor.opentelemetry import configure_azure_monitor
+
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
 env = Env()
@@ -64,7 +66,7 @@ INSTALLED_APPS = [
     'allauth',
     'allauth.account',
     'allauth.socialaccount',
-    'azure',
+    'azure_services',
     'rest_auth',
     'rest_auth.registration',
     'rest_framework',
